Remove category api call and all references. refactored into board games
