+++ .gitignore
@@ -1,1 +1,4 @@
 config/application.yml
+
+# Ignore application configuration
+/config/application.yml
