--- app/serializers/category_serializer.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class CategorySerializer
-  include JSONAPI::Serializer
-
-  set_id { nil }
-
-  attributes :id, :name, :url
-end
