--- app/services/category_service.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-class CategoryService
-  def self.conn
-    Faraday.new(
-      url: 'https://api.boardgameatlas.com/api/',
-      params: { client_id: ENV['CLIENT_ID'] }
-    )
-  end
-
-  def self.category
-    response = conn.get('game/categories') 
-    json = JSON.parse(response.body, symbolize_names: true)
-    json[:categories]
-  end
-end
