+++ spec/services/board_game_service_spec.rb
@@ -8,6 +8,26 @@ RSpec.describe 'BoardGamesService' do
   end
 
   describe 'class methods' do
+    describe '.category' do
+      it 'returns json data' do
+        categories = BoardGamesService.category
+
+        expect(categories).to be_an(Array)
+        expect(categories.first).to be_a(Hash)
+
+        expect(categories.count).to eq(122)
+
+        expect(categories.first).to have_key(:name)
+        expect(categories.first[:name]).to be_a(String)
+
+        expect(categories.first).to have_key(:id)
+        expect(categories.first[:id]).to be_a(String)
+
+        expect(categories.first).to have_key(:url)
+        expect(categories.first[:url]).to be_a(String)
+      end
+    end
+    
     describe '.find_board_games(name)' do
       it 'returns json data', :vcr do
         games = BoardGamesService.find_board_games('catan')
